moved complete test to other_test.go
